% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{attach_load_balancers}
\alias{attach_load_balancers}
\title{Attaches one or more Classic Load Balancers to the specified Auto Scaling group}
\usage{
attach_load_balancers(AutoScalingGroupName, LoadBalancerNames)
}
\arguments{
\item{AutoScalingGroupName}{The name of the Auto Scaling group.}

\item{LoadBalancerNames}{The names of the load balancers. You can specify up to 10 load balancers.}
}
\description{
Attaches one or more Classic Load Balancers to the specified Auto Scaling group.
}
\details{
To attach an Application Load Balancer instead, see AttachLoadBalancerTargetGroups.

To describe the load balancers for an Auto Scaling group, use DescribeLoadBalancers. To detach the load balancer from the Auto Scaling group, use DetachLoadBalancers.

For more information, see \href{http://docs.aws.amazon.com/autoscaling/ec2/userguide/attach-load-balancer-asg.html}{Attach a Load Balancer to Your Auto Scaling Group} in the \emph{Amazon EC2 Auto Scaling User Guide}.
}
\examples{
# This example attaches the specified load balancer to the specified Auto
# Scaling group.
attach_load_balancers(
  AutoScalingGroupName = "my-auto-scaling-group",
  LoadBalancerNames = list(
    "my-load-balancer"
  )
)

}
