% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{batch_delete_scheduled_action}
\alias{batch_delete_scheduled_action}
\title{Deletes one or more scheduled actions for the specified Auto Scaling group}
\usage{
batch_delete_scheduled_action(AutoScalingGroupName, ScheduledActionNames)
}
\arguments{
\item{AutoScalingGroupName}{The name of the Auto Scaling group.}

\item{ScheduledActionNames}{The names of the scheduled actions to delete. The maximum number allowed is 50.}
}
\description{
Deletes one or more scheduled actions for the specified Auto Scaling group.
}
\examples{

}
