% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{batch_put_scheduled_update_group_action}
\alias{batch_put_scheduled_update_group_action}
\title{Creates or updates one or more scheduled scaling actions for an Auto Scaling group}
\usage{
batch_put_scheduled_update_group_action(AutoScalingGroupName,
  ScheduledUpdateGroupActions)
}
\arguments{
\item{AutoScalingGroupName}{The name of the Auto Scaling group.}

\item{ScheduledUpdateGroupActions}{One or more scheduled actions. The maximum number allowed is 50.}
}
\description{
Creates or updates one or more scheduled scaling actions for an Auto Scaling group. When updating a scheduled scaling action, if you leave a parameter unspecified, the corresponding value remains unchanged.
}
\examples{

}
