% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{delete_launch_configuration}
\alias{delete_launch_configuration}
\title{Deletes the specified launch configuration}
\usage{
delete_launch_configuration(LaunchConfigurationName)
}
\arguments{
\item{LaunchConfigurationName}{[required] The name of the launch configuration.}
}
\description{
Deletes the specified launch configuration.
}
\details{
The launch configuration must not be attached to an Auto Scaling group. When this call completes, the launch configuration is no longer available for use.
}
\section{Accepted Parameters}{
\preformatted{delete_launch_configuration(
  LaunchConfigurationName = "string"
)
}
}

\examples{
# This example deletes the specified launch configuration.
\donttest{delete_launch_configuration(
  LaunchConfigurationName = "my-launch-config"
)}

}
