% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{describe_account_limits}
\alias{describe_account_limits}
\title{Describes the current Auto Scaling resource limits for your AWS account}
\usage{
describe_account_limits()
}
\description{
Describes the current Auto Scaling resource limits for your AWS account.
}
\details{
For information about requesting an increase in these limits, see \href{http://docs.aws.amazon.com/autoscaling/ec2/userguide/as-account-limits.html}{Auto Scaling Limits} in the \emph{Amazon EC2 Auto Scaling User Guide}.
}
\examples{
# This example describes the Auto Scaling limits for your AWS account.
describe_account_limits()

}
