% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{describe_adjustment_types}
\alias{describe_adjustment_types}
\title{Describes the policy adjustment types for use with PutScalingPolicy}
\usage{
describe_adjustment_types()
}
\description{
Describes the policy adjustment types for use with PutScalingPolicy.
}
\examples{
# This example describes the available adjustment types.
\donttest{describe_adjustment_types()}

}
