% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{describe_auto_scaling_groups}
\alias{describe_auto_scaling_groups}
\title{Describes one or more Auto Scaling groups}
\usage{
describe_auto_scaling_groups(AutoScalingGroupNames = NULL, NextToken = NULL,
  MaxRecords = NULL)
}
\arguments{
\item{AutoScalingGroupNames}{The names of the Auto Scaling groups. You can specify up to \code{MaxRecords} names. If you omit this parameter, all Auto Scaling groups are described.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}

\item{MaxRecords}{The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.}
}
\description{
Describes one or more Auto Scaling groups.
}
\examples{
# This example describes the specified Auto Scaling group.
describe_auto_scaling_groups(
  AutoScalingGroupNames = list(
    "my-auto-scaling-group"
  )
)

}
