% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{describe_auto_scaling_instances}
\alias{describe_auto_scaling_instances}
\title{Describes one or more Auto Scaling instances}
\usage{
describe_auto_scaling_instances(InstanceIds = NULL, MaxRecords = NULL,
  NextToken = NULL)
}
\arguments{
\item{InstanceIds}{The IDs of the instances. You can specify up to \code{MaxRecords} IDs. If you omit this parameter, all Auto Scaling instances are described. If you specify an ID that does not exist, it is ignored with no error.}

\item{MaxRecords}{The maximum number of items to return with this call. The default value is 50 and the maximum value is 50.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
Describes one or more Auto Scaling instances.
}
\examples{
# This example describes the specified Auto Scaling instance.
describe_auto_scaling_instances(
  InstanceIds = list(
    "i-4ba0837f"
  )
)

}
