% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{describe_auto_scaling_notification_types}
\alias{describe_auto_scaling_notification_types}
\title{Describes the notification types that are supported by Amazon EC2 Auto Scaling}
\usage{
describe_auto_scaling_notification_types()
}
\description{
Describes the notification types that are supported by Amazon EC2 Auto Scaling.
}
\examples{
# This example describes the available notification types.
\donttest{describe_auto_scaling_notification_types()}

}
