% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{describe_launch_configurations}
\alias{describe_launch_configurations}
\title{Describes one or more launch configurations}
\usage{
describe_launch_configurations(LaunchConfigurationNames = NULL,
  NextToken = NULL, MaxRecords = NULL)
}
\arguments{
\item{LaunchConfigurationNames}{The launch configuration names. If you omit this parameter, all launch configurations are described.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}

\item{MaxRecords}{The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.}
}
\description{
Describes one or more launch configurations.
}
\section{Accepted Parameters}{
\preformatted{describe_launch_configurations(
  LaunchConfigurationNames = list(
    "string"
  ),
  NextToken = "string",
  MaxRecords = 123
)
}
}

\examples{
# This example describes the specified launch configuration.
\donttest{describe_launch_configurations(
  LaunchConfigurationNames = list(
    "my-launch-config"
  )
)}

}
