% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{describe_lifecycle_hooks}
\alias{describe_lifecycle_hooks}
\title{Describes the lifecycle hooks for the specified Auto Scaling group}
\usage{
describe_lifecycle_hooks(
  AutoScalingGroupName = "string",
  LifecycleHookNames = list(
    "string"
  )
)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{LifecycleHookNames}{The names of one or more lifecycle hooks. If you omit this parameter, all lifecycle hooks are described.}
}
\description{
Describes the lifecycle hooks for the specified Auto Scaling group.
}
\examples{
# This example describes the lifecycle hooks for the specified Auto
# Scaling group.
describe_lifecycle_hooks(
  AutoScalingGroupName = "my-auto-scaling-group"
)

}
