% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{describe_load_balancer_target_groups}
\alias{describe_load_balancer_target_groups}
\title{Describes the target groups for the specified Auto Scaling group}
\usage{
describe_load_balancer_target_groups(AutoScalingGroupName, NextToken = NULL,
  MaxRecords = NULL)
}
\arguments{
\item{AutoScalingGroupName}{The name of the Auto Scaling group.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}

\item{MaxRecords}{The maximum number of items to return with this call. The default value is 100 and the maximum value is 100.}
}
\description{
Describes the target groups for the specified Auto Scaling group.
}
\examples{
# This example describes the target groups attached to the specified Auto
# Scaling group.
describe_load_balancer_target_groups(
  AutoScalingGroupName = "my-auto-scaling-group"
)

}
