% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{describe_metric_collection_types}
\alias{describe_metric_collection_types}
\title{Describes the available CloudWatch metrics for Amazon EC2 Auto Scaling}
\usage{
describe_metric_collection_types()
}
\description{
Describes the available CloudWatch metrics for Amazon EC2 Auto Scaling.
}
\details{
Note that the \code{GroupStandbyInstances} metric is not returned by default. You must explicitly request this metric when calling EnableMetricsCollection.
}
\examples{
# This example describes the available metric collection types.
\donttest{describe_metric_collection_types()}

}
