% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{describe_scaling_process_types}
\alias{describe_scaling_process_types}
\title{Describes the scaling process types for use with ResumeProcesses and SuspendProcesses}
\usage{
describe_scaling_process_types()
}
\description{
Describes the scaling process types for use with ResumeProcesses and SuspendProcesses.
}
\examples{
# This example describes the Auto Scaling process types.
\donttest{describe_scaling_process_types()}

}
