% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{describe_termination_policy_types}
\alias{describe_termination_policy_types}
\title{Describes the termination policies supported by Amazon EC2 Auto Scaling}
\usage{
describe_termination_policy_types()
}
\description{
Describes the termination policies supported by Amazon EC2 Auto Scaling.
}
\examples{
# This example describes the available termination policy types.
describe_termination_policy_types()

}
