% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{detach_load_balancer_target_groups}
\alias{detach_load_balancer_target_groups}
\title{Detaches one or more target groups from the specified Auto Scaling group}
\usage{
detach_load_balancer_target_groups(AutoScalingGroupName, TargetGroupARNs)
}
\arguments{
\item{AutoScalingGroupName}{The name of the Auto Scaling group.}

\item{TargetGroupARNs}{The Amazon Resource Names (ARN) of the target groups. You can specify up to 10 target groups.}
}
\description{
Detaches one or more target groups from the specified Auto Scaling group.
}
\examples{
# This example detaches the specified target group from the specified Auto
# Scaling group
detach_load_balancer_target_groups(
  AutoScalingGroupName = "my-auto-scaling-group",
  TargetGroupARNs = list(
    "arn:aws:elasticloadbalancing:us-west-2:123456789012:targetgroup/my-targets/73e2d6bc24d8a067"
  )
)

}
