% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{detach_load_balancers}
\alias{detach_load_balancers}
\title{Detaches one or more Classic Load Balancers from the specified Auto Scaling group}
\usage{
detach_load_balancers(AutoScalingGroupName, LoadBalancerNames)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{LoadBalancerNames}{[required] The names of the load balancers. You can specify up to 10 load balancers.}
}
\description{
Detaches one or more Classic Load Balancers from the specified Auto Scaling group.
}
\details{
Note that this operation detaches only Classic Load Balancers. If you have Application Load Balancers, use DetachLoadBalancerTargetGroups instead.

When you detach a load balancer, it enters the \code{Removing} state while deregistering the instances in the group. When all instances are deregistered, then you can no longer describe the load balancer using DescribeLoadBalancers. Note that the instances remain running.
}
\section{Accepted Parameters}{
\preformatted{detach_load_balancers(
  AutoScalingGroupName = "string",
  LoadBalancerNames = list(
    "string"
  )
)
}
}

\examples{
# This example detaches the specified load balancer from the specified
# Auto Scaling group.
\donttest{detach_load_balancers(
  AutoScalingGroupName = "my-auto-scaling-group",
  LoadBalancerNames = list(
    "my-load-balancer"
  )
)}

}
