% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{enter_standby}
\alias{enter_standby}
\title{Moves the specified instances into the standby state}
\usage{
enter_standby(AutoScalingGroupName, ShouldDecrementDesiredCapacity,
  InstanceIds = NULL)
}
\arguments{
\item{AutoScalingGroupName}{The name of the Auto Scaling group.}

\item{ShouldDecrementDesiredCapacity}{Indicates whether to decrement the desired capacity of the Auto Scaling group by the number of instances moved to \code{Standby} mode.}

\item{InstanceIds}{The IDs of the instances. You can specify up to 20 instances.}
}
\description{
Moves the specified instances into the standby state.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/autoscaling/ec2/userguide/as-enter-exit-standby.html}{Temporarily Removing Instances from Your Auto Scaling Group} in the \emph{Amazon EC2 Auto Scaling User Guide}.
}
\examples{
# This example puts the specified instance into standby mode.
enter_standby(
  AutoScalingGroupName = "my-auto-scaling-group",
  InstanceIds = list(
    "i-93633f9b"
  ),
  ShouldDecrementDesiredCapacity = TRUE
)

}
