% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{exit_standby}
\alias{exit_standby}
\title{Moves the specified instances out of the standby state}
\usage{
exit_standby(AutoScalingGroupName, InstanceIds = NULL)
}
\arguments{
\item{AutoScalingGroupName}{The name of the Auto Scaling group.}

\item{InstanceIds}{The IDs of the instances. You can specify up to 20 instances.}
}
\description{
Moves the specified instances out of the standby state.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/autoscaling/ec2/userguide/as-enter-exit-standby.html}{Temporarily Removing Instances from Your Auto Scaling Group} in the \emph{Amazon EC2 Auto Scaling User Guide}.
}
\examples{
# This example moves the specified instance out of standby mode.
exit_standby(
  AutoScalingGroupName = "my-auto-scaling-group",
  InstanceIds = list(
    "i-93633f9b"
  )
)

}
