% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_package.R
\docType{package}
\name{paws.autoscaling-package}
\alias{paws.autoscaling}
\alias{paws.autoscaling-package}
\title{paws.autoscaling: Auto Scaling}
\description{
Amazon EC2 Auto Scaling is designed to
    automatically launch or terminate EC2 instances based on user-defined
    policies, schedules, and health checks. Use this service with AWS Auto
    Scaling, Amazon CloudWatch, and Elastic Load Balancing.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
