% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{put_lifecycle_hook}
\alias{put_lifecycle_hook}
\title{Creates or updates a lifecycle hook for the specified Auto Scaling Group}
\usage{
put_lifecycle_hook(LifecycleHookName, AutoScalingGroupName,
  LifecycleTransition = NULL, RoleARN = NULL,
  NotificationTargetARN = NULL, NotificationMetadata = NULL,
  HeartbeatTimeout = NULL, DefaultResult = NULL)
}
\arguments{
\item{LifecycleHookName}{The name of the lifecycle hook.}

\item{AutoScalingGroupName}{The name of the Auto Scaling group.}

\item{LifecycleTransition}{The instance state to which you want to attach the lifecycle hook. The possible values are:
\itemize{
\item autoscaling:EC2\_INSTANCE\_LAUNCHING
\item autoscaling:EC2\_INSTANCE\_TERMINATING
}

This parameter is required for new lifecycle hooks, but optional when updating existing hooks.}

\item{RoleARN}{The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target.

This parameter is required for new lifecycle hooks, but optional when updating existing hooks.}

\item{NotificationTargetARN}{The ARN of the notification target that Amazon EC2 Auto Scaling will use to notify you when an instance is in the transition state for the lifecycle hook. This target can be either an SQS queue or an SNS topic. If you specify an empty string, this overrides the current ARN.

This operation uses the JSON format when sending notifications to an Amazon SQS queue, and an email key/value pair format when sending notifications to an Amazon SNS topic.

When you specify a notification target, Amazon EC2 Auto Scaling sends it a test message. Test messages contains the following additional key/value pair: \code{"Event": "autoscaling:TEST_NOTIFICATION"}.}

\item{NotificationMetadata}{Contains additional information that you want to include any time Amazon EC2 Auto Scaling sends a message to the notification target.}

\item{HeartbeatTimeout}{The maximum time, in seconds, that can elapse before the lifecycle hook times out. The range is from 30 to 7200 seconds. The default is 3600 seconds (1 hour).

If the lifecycle hook times out, Amazon EC2 Auto Scaling performs the default action. You can prevent the lifecycle hook from timing out by calling RecordLifecycleActionHeartbeat.}

\item{DefaultResult}{Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. This parameter can be either \code{CONTINUE} or \code{ABANDON}. The default value is \code{ABANDON}.}
}
\description{
Creates or updates a lifecycle hook for the specified Auto Scaling Group.
}
\details{
A lifecycle hook tells Amazon EC2 Auto Scaling that you want to perform an action on an instance that is not actively in service; for example, either when the instance launches or before the instance terminates.

This step is a part of the procedure for adding a lifecycle hook to an Auto Scaling group:
\enumerate{
\item (Optional) Create a Lambda function and a rule that allows CloudWatch Events to invoke your Lambda function when Amazon EC2 Auto Scaling launches or terminates instances.
\item (Optional) Create a notification target and an IAM role. The target can be either an Amazon SQS queue or an Amazon SNS topic. The role allows Amazon EC2 Auto Scaling to publish lifecycle notifications to the target.
\item \strong{Create the lifecycle hook. Specify whether the hook is used when the instances launch or terminate.}
\item If you need more time, record the lifecycle action heartbeat to keep the instance in a pending state.
\item If you finish before the timeout period ends, complete the lifecycle action.
}

For more information, see \href{http://docs.aws.amazon.com/autoscaling/ec2/userguide/lifecycle-hooks.html}{Auto Scaling Lifecycle Hooks} in the \emph{Amazon EC2 Auto Scaling User Guide}.

If you exceed your maximum limit of lifecycle hooks, which by default is 50 per Auto Scaling group, the call fails. For information about updating this limit, see \href{http://docs.aws.amazon.com/general/latest/gr/aws_service_limits.html}{AWS Service Limits} in the \emph{Amazon Web Services General Reference}.
}
\examples{
# This example creates a lifecycle hook.
put_lifecycle_hook(
  AutoScalingGroupName = "my-auto-scaling-group",
  LifecycleHookName = "my-lifecycle-hook",
  LifecycleTransition = "autoscaling:EC2_INSTANCE_LAUNCHING",
  NotificationTargetARN = "arn:aws:sns:us-west-2:123456789012:my-sns-topic --role-arn",
  RoleARN = "arn:aws:iam::123456789012:role/my-auto-scaling-role"
)

}
