% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{put_scheduled_update_group_action}
\alias{put_scheduled_update_group_action}
\title{Creates or updates a scheduled scaling action for an Auto Scaling group}
\usage{
put_scheduled_update_group_action(AutoScalingGroupName,
  ScheduledActionName, Time = NULL, StartTime = NULL, EndTime = NULL,
  Recurrence = NULL, MinSize = NULL, MaxSize = NULL,
  DesiredCapacity = NULL)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{ScheduledActionName}{[required] The name of this scaling action.}

\item{Time}{This parameter is deprecated.}

\item{StartTime}{The time for this action to start, in "YYYY-MM-DDThh:mm:ssZ" format in UTC/GMT only (for example, \code{2014-06-01T00:00:00Z}).

If you specify \code{Recurrence} and \code{StartTime}, Amazon EC2 Auto Scaling performs the action at this time, and then performs the action based on the specified recurrence.

If you try to schedule your action in the past, Amazon EC2 Auto Scaling returns an error message.}

\item{EndTime}{The time for the recurring schedule to end. Amazon EC2 Auto Scaling does not perform the action after this time.}

\item{Recurrence}{The recurring schedule for this action, in Unix cron syntax format. For more information about this format, see \href{http://crontab.org}{Crontab}.}

\item{MinSize}{The minimum size for the Auto Scaling group.}

\item{MaxSize}{The maximum size for the Auto Scaling group.}

\item{DesiredCapacity}{The number of EC2 instances that should be running in the group.}
}
\description{
Creates or updates a scheduled scaling action for an Auto Scaling group. If you leave a parameter unspecified when updating a scheduled scaling action, the corresponding value remains unchanged.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/autoscaling/ec2/userguide/schedule_time.html}{Scheduled Scaling} in the \emph{Amazon EC2 Auto Scaling User Guide}.
}
\section{Accepted Parameters}{
\preformatted{put_scheduled_update_group_action(
  AutoScalingGroupName = "string",
  ScheduledActionName = "string",
  Time = as.POSIXct("2015-01-01"),
  StartTime = as.POSIXct("2015-01-01"),
  EndTime = as.POSIXct("2015-01-01"),
  Recurrence = "string",
  MinSize = 123,
  MaxSize = 123,
  DesiredCapacity = 123
)
}
}

\examples{
# This example adds the specified scheduled action to the specified Auto
# Scaling group.
\donttest{put_scheduled_update_group_action(
  AutoScalingGroupName = "my-auto-scaling-group",
  DesiredCapacity = 4L,
  EndTime = "2014-05-12T08:00:00Z",
  MaxSize = 6L,
  MinSize = 2L,
  ScheduledActionName = "my-scheduled-action",
  StartTime = "2014-05-12T08:00:00Z"
)}

}
