% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{resume_processes}
\alias{resume_processes}
\title{Resumes the specified suspended automatic scaling processes, or all suspended process, for the specified Auto Scaling group}
\usage{
resume_processes(
  AutoScalingGroupName = "string",
  ScalingProcesses = list(
    "string"
  )
)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{ScalingProcesses}{One or more of the following processes. If you omit this parameter, all processes are specified.
\itemize{
\item \code{Launch}
\item \code{Terminate}
\item \code{HealthCheck}
\item \code{ReplaceUnhealthy}
\item \code{AZRebalance}
\item \code{AlarmNotification}
\item \code{ScheduledActions}
\item \code{AddToLoadBalancer}
}}
}
\description{
Resumes the specified suspended automatic scaling processes, or all suspended process, for the specified Auto Scaling group.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/autoscaling/ec2/userguide/as-suspend-resume-processes.html}{Suspending and Resuming Scaling Processes} in the \emph{Amazon EC2 Auto Scaling User Guide}.
}
\examples{
# This example resumes the specified suspended scaling process for the
# specified Auto Scaling group.
resume_processes(
  AutoScalingGroupName = "my-auto-scaling-group",
  ScalingProcesses = list(
    "AlarmNotification"
  )
)

}
