% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{set_desired_capacity}
\alias{set_desired_capacity}
\title{Sets the size of the specified Auto Scaling group}
\usage{
set_desired_capacity(
  AutoScalingGroupName = "string",
  DesiredCapacity = 123,
  HonorCooldown = TRUE|FALSE
)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{DesiredCapacity}{[required] The number of EC2 instances that should be running in the Auto Scaling group.}

\item{HonorCooldown}{Indicates whether Amazon EC2 Auto Scaling waits for the cooldown period to complete before initiating a scaling activity to set your Auto Scaling group to its new capacity. By default, Amazon EC2 Auto Scaling does not honor the cooldown period during manual scaling activities.}
}
\description{
Sets the size of the specified Auto Scaling group.
}
\details{
For more information about desired capacity, see \href{http://docs.aws.amazon.com/autoscaling/ec2/userguide/WhatIsAutoScaling.html}{What Is Amazon EC2 Auto Scaling?} in the \emph{Amazon EC2 Auto Scaling User Guide}.
}
\examples{
# This example sets the desired capacity for the specified Auto Scaling
# group.
set_desired_capacity(
  AutoScalingGroupName = "my-auto-scaling-group",
  DesiredCapacity = 2L,
  HonorCooldown = TRUE
)

}
