% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{delete_lifecycle_hook}
\alias{delete_lifecycle_hook}
\title{Deletes the specified lifecycle hook}
\usage{
delete_lifecycle_hook(LifecycleHookName, AutoScalingGroupName)
}
\arguments{
\item{LifecycleHookName}{The name of the lifecycle hook.}

\item{AutoScalingGroupName}{The name of the Auto Scaling group.}
}
\description{
Deletes the specified lifecycle hook.
}
\details{
If there are any outstanding lifecycle actions, they are completed first (\code{ABANDON} for launching instances, \code{CONTINUE} for terminating instances).
}
\examples{
# This example deletes the specified lifecycle hook.
delete_lifecycle_hook(
  AutoScalingGroupName = "my-auto-scaling-group",
  LifecycleHookName = "my-lifecycle-hook"
)

}
