% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{delete_notification_configuration}
\alias{delete_notification_configuration}
\title{Deletes the specified notification}
\usage{
delete_notification_configuration(
  AutoScalingGroupName = "string",
  TopicARN = "string"
)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{TopicARN}{[required] The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (Amazon SNS) topic.}
}
\description{
Deletes the specified notification.
}
\examples{
# This example deletes the specified notification from the specified Auto
# Scaling group.
delete_notification_configuration(
  AutoScalingGroupName = "my-auto-scaling-group",
  TopicARN = "arn:aws:sns:us-west-2:123456789012:my-sns-topic"
)

}
