% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{delete_policy}
\alias{delete_policy}
\title{Deletes the specified Auto Scaling policy}
\usage{
delete_policy(AutoScalingGroupName = NULL, PolicyName)
}
\arguments{
\item{AutoScalingGroupName}{The name of the Auto Scaling group.}

\item{PolicyName}{[required] The name or Amazon Resource Name (ARN) of the policy.}
}
\description{
Deletes the specified Auto Scaling policy.
}
\details{
Deleting a policy deletes the underlying alarm action, but does not delete the alarm, even if it no longer has an associated action.
}
\section{Accepted Parameters}{
\preformatted{delete_policy(
  AutoScalingGroupName = "string",
  PolicyName = "string"
)
}
}

\examples{
# This example deletes the specified Auto Scaling policy.
\donttest{delete_policy(
  AutoScalingGroupName = "my-auto-scaling-group",
  PolicyName = "ScaleIn"
)}

}
