% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{delete_scheduled_action}
\alias{delete_scheduled_action}
\title{Deletes the specified scheduled action}
\usage{
delete_scheduled_action(AutoScalingGroupName, ScheduledActionName)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{ScheduledActionName}{[required] The name of the action to delete.}
}
\description{
Deletes the specified scheduled action.
}
\section{Accepted Parameters}{
\preformatted{delete_scheduled_action(
  AutoScalingGroupName = "string",
  ScheduledActionName = "string"
)
}
}

\examples{
# This example deletes the specified scheduled action from the specified
# Auto Scaling group.
\donttest{delete_scheduled_action(
  AutoScalingGroupName = "my-auto-scaling-group",
  ScheduledActionName = "my-scheduled-action"
)}

}
