% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.autoscaling_operations.R
\name{delete_tags}
\alias{delete_tags}
\title{Deletes the specified tags}
\usage{
delete_tags(Tags)
}
\arguments{
\item{Tags}{One or more tags.}
}
\description{
Deletes the specified tags.
}
\examples{
# This example deletes the specified tag from the specified Auto Scaling
# group.
delete_tags(
  Tags = list(
    list(
      Key = "Dept",
      ResourceId = "my-auto-scaling-group",
      ResourceType = "auto-scaling-group",
      Value = "Research"
    )
  )
)

}
